requirements.txt
# Healthcare MLOps System - Requirements
# Python 3.11+

# Core ML
scikit-learn==1.4.0
xgboost==2.0.3
optuna==3.5.0
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
joblib==1.3.2
imbalanced-learn==0.12.0

# MLflow
mlflow==2.10.0
boto3==1.34.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Monitoring
prometheus-client==0.19.0

# Airflow
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.17.0

# Data validation
great-expectations==0.18.8

# Docker / Cloud
docker==7.0.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# Code quality
black==24.1.0
flake8==7.0.0
isort==5.13.2
